Move mockSetCartEndpoint/mockGetCartEndpointWith into test
