+++ client/my-sites/checkout/composite-checkout/wpcom/index.js
@@ -5,7 +5,6 @@ import WPCheckout from './components/wp-checkout';
 import WPCheckoutErrorBoundary from './components/wp-checkout-error-boundary';
 import { useShoppingCart } from './hooks/use-shopping-cart';
 import { useWpcomStore } from './hooks/wpcom-store';
-import { mockSetCartEndpoint, mockGetCartEndpointWith } from './mock/cart-endpoint';
 import FormFieldAnnotation from './components/form-field-annotation';
 import { getNonProductWPCOMCartItemTypes } from './lib/translate-cart';
 import { areDomainsInLineItems } from './hooks/has-domains';
@@ -26,8 +25,6 @@ export {
 	WPCheckoutErrorBoundary,
 	useShoppingCart,
 	useWpcomStore,
-	mockSetCartEndpoint,
-	mockGetCartEndpointWith,
 	FormFieldAnnotation,
 	WPCOMCartItem,
 	CheckoutCartItem,
